requirements.txt
altgraph==0.17
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
future==0.18.2
idna==2.10
importlib-metadata==3.4.0
lxml==4.5.1
pefile==2019.4.18
Pillow==8.1.0
pyee==7.0.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyppeteer2==0.2.2
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.15.2.3.0.1
python-dotenv==0.15.0
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==8.1
zipp==3.4.0
